+++ src/main/java/net/watchbox/domain/box/facade/shared/SharedBoxContentFacade.java
@@ -128,7 +128,7 @@ public class SharedBoxContentFacade {
         // 5. 응답
         return ContentPageResponse.builder()
                 .contentItemList(contentItemList)
-                .totalCount(contentItemList.size())
+                .totalCount((long) contentItemList.size())
 //                .totalPages()
 //                .currentPage()
                 .build();
